Changed folders' structure and renamed files
