--- app/app/nested/page.jsx
@@ -1,5 +1,0 @@
-import TestPage from "@/test/TestPage";
-
-export default function NestedPage () {
-    return <TestPage />
-}
